resume_agent/requirements.txt
streamlit
openai
sentence-transformers
faiss-cpu
pdfplumber
python-dotenv
numpy
pandas
